[fromtree][NXP][SE05x] SE05x Crypto Host fallback updated to use PSA APIs for NXP MCUs (#72831)
* SE05x Crypto Host fallback updated to use PSA APIs for NXP MCUs

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------


(cherry picked from commit 9d45993)

+++ src/platform/nxp/crypto/se05x/CHIPCryptoPAL_se05x.h
@@ -31,7 +31,7 @@ typedef struct hsm_pake_context_s
 {
     uint8_t spake_context[32];
     size_t spake_context_len;
-    uint8_t spake_objId;
+    uint16_t spake_objId;
 } hsm_pake_context_t;
 #endif // #if ((ENABLE_SE05X_SPAKE_VERIFIER) || (ENABLE_SE05X_SPAKE_PROVER))
